requirements.txt
torch
torch-geometric
numpy
pandas
scikit-learn
scipy
matplotlib
biopandas
tqdm
fair-esm
requests
